+++ client/src/app/Pixi/GameBoard/ConfirmCancelButtons.ts
@@ -3,9 +3,9 @@ import { PixiManager, GameZIndex } from '../../../api/PixiManager';
 import { PieceType } from '../../../_types/global/GlobalTypes';
 import { CHAR_H, LINE_SPACING } from '../Utils/FontLoader';
 import { PixiObject } from '../PixiObject';
-import { LinkObject, TextAlign } from '../Text';
 import { GameGrid } from './GameGrid';
 import { GameBoardObject } from './GridObject';
+import { LinkObject, TextAlign } from '../Utils/Text';
 
 // TODO make a GridObject class?
 export class ConfirmCancelButtons extends GameBoardObject {
