+++ client/src/app/Pixi/PixiObject.ts
@@ -1,10 +1,8 @@
 import autoBind from 'auto-bind';
 import * as PIXI from 'pixi.js';
 import { GameZIndex, PixiManager } from '../../api/PixiManager';
-import { ZKPiece } from '../../_types/global/GlobalTypes';
-import Game from '../Game';
 import { CanvasCoords } from './@PixiTypes';
-import { objFromHitArea } from './PixiUtils';
+import { objFromHitArea } from './Utils/PixiUtils';
 
 const autoIncrement = (() => {
   let count = 0;
