+++ client/src/app/Pixi/ResourceBars.ts
@@ -1,9 +1,9 @@
 import { GameZIndex, PixiManager } from '../../api/PixiManager';
 import { PixiObject } from './PixiObject';
 import * as PIXI from 'pixi.js';
-import { BASELINE_ICONS, BASELINE_TEXT, ICONS } from './TextureLoader';
 import { CanvasCoords } from './@PixiTypes';
-import { TextObject } from './Text';
+import { TextObject } from './Utils/Text';
+import { BASELINE_TEXT, BASELINE_ICONS, ICONS } from './Utils/TextureLoader';
 
 const LABEL_WIDTH = 32; // width of 'Gold:'
 const LABEL_M_RIGHT = 2; // right margin
