+++ client/src/app/Pixi/Utils/PixiUtils.ts
@@ -1,7 +1,6 @@
 import * as PIXI from 'pixi.js';
-import { PixiManager } from '../../api/PixiManager';
-import { BoardLocation } from '../../_types/global/GlobalTypes';
-import { BoardCoords, CanvasCoords, LineAlignment } from './@PixiTypes';
+import { BoardLocation } from '../../../_types/global/GlobalTypes';
+import { LineAlignment, BoardCoords, CanvasCoords } from '../@PixiTypes';
 
 // general-purpose, smaller utils
