+++ client/src/app/Pixi/Utils/StatIcon.ts
@@ -1,6 +1,6 @@
 import * as PIXI from 'pixi.js';
-import { PixiManager } from '../../api/PixiManager';
-import { PixiObject } from './PixiObject';
+import { PixiManager } from '../../../api/PixiManager';
+import { PixiObject } from '../PixiObject';
 import { TextAlign, TextObject } from './Text';
 import { getBombSprite, getHeartSprite } from './TextureLoader';
