+++ client/src/app/Pixi/Utils/TextureLoader.ts
